+++ NEWS.md
@@ -1,3 +1,7 @@
+# 
+ - Add decisionpt gadget
+
+
 # Version 0.33.0 (2017-11-01)
  - Factor out randomForest specific functions parRandomForest and varImpPlot2
  - Moved rf_test_names -> check_features